+++ Saitynu API/Saitynu API/Controllers/PlayersController.cs
@@ -8,6 +8,8 @@ using Saitynu_API.Data.Repositories;
 using Saitynu_API.Data.Entities;
 using Saitynu_API.Data.Dtos.Players;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
+using Saitynu_API.Auth.Model;
 
 namespace Saitynu_API.Controllers
 {
@@ -17,33 +19,51 @@ namespace Saitynu_API.Controllers
     {
         private readonly IPlayersRepository _playersRepository;
         private readonly IMapper _mapper;
+        private readonly IAuthorizationService _authorizationService;
 
-        public PlayersController(IPlayersRepository playersRepository, IMapper mapper)
+        public PlayersController(IPlayersRepository playersRepository, IMapper mapper, IAuthorizationService authorizationService)
         {
             _playersRepository = playersRepository;
             _mapper = mapper;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
+        [Authorize(Roles = "Admin, Tester")]
         public async Task<IEnumerable<PlayerDto>> GetAll()
         {
-            return (await _playersRepository.GetAll()).Select(o => _mapper.Map<PlayerDto>(o));
+            //return (await _playersRepository.GetAll()).Select(o => _mapper.Map<PlayerDto>(o));
+            var players = await _playersRepository.GetAll();
+            return players.Select(o => _mapper.Map<PlayerDto>(o));
         }
 
         [HttpGet("{id}")]
+        [Authorize(Roles = "Admin, Tester, Player")]
         public async Task<ActionResult<PlayerDto>> Get(int id)
         {
             var player = await _playersRepository.Get(id);
             if (player == null)
                 return NotFound("Player with id " + id + " not found.");
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, player, PolicyNames.SameUser);
+            if (!authorizationResult.Succeeded)
+                return Forbid();
+
             return Ok(_mapper.Map<PlayerDto>(player));
         }
 
         [HttpPost]
+        [Authorize(Roles = "Admin, Tester, Player")]
         public async Task<ActionResult<PlayerDto>> Post(CreatePlayerDto playerDto)
         {
+            var players = await _playersRepository.GetAll();
+            var chosenPlayers = players.Where(o => o.UserId == User.FindFirst(CustomClaims.UserId).Value).Select(o => o.UserId);
+            if (chosenPlayers.Count() > 0 && User.IsInRole(RestUserRoles.Player))
+                return BadRequest("This user (player role) already has an existing player");
+
             var player = _mapper.Map<Player>(playerDto);
+            var id = User.FindFirst("userId").Value;
+            player.UserId = id;
 
             await _playersRepository.Create(player);
 
@@ -51,11 +71,16 @@ namespace Saitynu_API.Controllers
         }
 
         [HttpPut("{id}")]
+        [Authorize(Roles = "Admin, Tester, Player")]
         public async Task<ActionResult<PlayerDto>> Put(int id, UpdatePlayerDto playerDto)
         {
             var player = await _playersRepository.Get(id);
             if (player == null)
                 return NotFound("Player with id " + id + " not found.");
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, player, PolicyNames.SameUser);
+            if (!authorizationResult.Succeeded)
+                return Forbid();
 
             //player.Nick = playerDto.Nick; jei tik viena keiciam kintamaji
             _mapper.Map(playerDto, player);
@@ -66,11 +91,16 @@ namespace Saitynu_API.Controllers
         }
 
         [HttpDelete("{id}")]
+        [Authorize(Roles = "Admin, Tester, Player")]
         public async Task<ActionResult<PlayerDto>> Delete(int id)
         {
             var player = await _playersRepository.Get(id);
             if (player == null)
                 return NotFound("Player with id " + id + " not found.");
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, player, PolicyNames.SameUser);
+            if (!authorizationResult.Succeeded)
+                return Forbid();
 
             await _playersRepository.Delete(player);
